Fix issues

+++ project/src/depth_image.cpp
@@ -22,7 +22,7 @@ std::tuple<double, double> DepthImage::compareTo(const DepthImage& o) {
     double s = histogram.minimumSum(o.histogram);
     double sum = 0;
     auto& otherData = o.pixelData;
-    for(int i = 0; i < pixelData.size(); i++) {
+    for(size_t i = 0; i < pixelData.size(); i++) {
         auto diff = (pixelData[i] - otherData[i]);
         auto square = diff * diff;
         sum += square;
